+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add performance test for single module detection
 * Pandas methods are now (all) available from `ImageSequence.pandas`
+* `Image` now supports lazy loading of data
 
 ## 0.2.3